Update footer.html
Added required alt attributes and descriptions for all the images (required by accessibility devices like screen readers).
Added Google+ page link. The text can be replaced with an image later on.
Updated the copyright so that is correctly shows that all work on the site from 2012 to 2014 falls under it.